+++ src/felupe/assembly/__init__.py
@@ -1,7 +1,7 @@
-from ._form import Form
+from . import expression
 from ._integral import IntegralForm
 
 __all__ = [
     "IntegralForm",
-    "Form",
+    "expression",
 ]