+++ src/felupe/assembly/expression/__init__.py
@@ -1,0 +1,7 @@
+from ._basis import Basis
+from ._decorator import FormExpressionDecorator as Form
+
+__all__ = [
+    "Basis",
+    "Form",
+]