+++ src/felupe/mechanics/_solidbody_pressure.py
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from .._assembly import IntegralForm
+from ..assembly import IntegralForm
 from ..constitution import AreaChange
 from ._helpers import Assemble, Results